+++ src/invoke-lambda/exec/spawn.js
@@ -5,28 +5,37 @@ let errors = require('../../lib/errors')
 let { invocations } = require('../../arc/_runtime-api')
 
 module.exports = function spawnChild (params, callback) {
-  let { args, context, command, lambda, options, requestID, timeout } = params
+  let { args, coldstart, context, command, lambda, options, requestID, timeout } = params
   let { apiType, update } = context
   let isInLambda = process.env.AWS_LAMBDA_FUNCTION_NAME
   let timedOut = false
 
   // Let's go!
-  let child = spawn(command, args, options)
-  let pid = child.pid
-  let error
-  let closed
+  let pid = 'init'
+  let child, error, closed
+  function start () {
+    child = spawn(command, args, options)
+    pid = child.pid
 
-  child.stdout.on('data', data => process.stdout.write('\n' + data))
-  child.stderr.on('data', data => process.stderr.write('\n' + data))
-  child.on('error', err => {
-    error = err
-    // Seen some non-string oob errors come via binary compilation
-    if (err.code) shutdown('error')
-  })
-  child.on('close', (code, signal) => {
-    update.debug.status(`[${requestID}] Emitted 'close' (pid ${pid}, code '${code}', signal '${signal}')`)
-    shutdown('child process closure')
-  })
+    child.stdout.on('data', data => process.stdout.write('\n' + data))
+    child.stderr.on('data', data => process.stderr.write('\n' + data))
+    child.on('error', err => {
+      error = err
+      // Seen some non-string oob errors come via binary compilation
+      if (err.code) shutdown('error')
+    })
+    child.on('close', (code, signal) => {
+      update.debug.status(`[${requestID}] Emitted 'close' (pid ${pid}, code '${code}', signal '${signal}')`)
+      shutdown('child process closure')
+    })
+  }
+  if (coldstart) {
+    if (coldstart < timeout) setTimeout(start, coldstart)
+    else {
+      update.debug.status(`[${requestID}] Coldstart simulator: coldstart of ${coldstart}ms exceeds timeout of ${timeout}ms, not spawning the Lambda`)
+    }
+  }
+  else start()
 
   // Set an execution timeout
   let to = setTimeout(function () {
@@ -63,7 +72,7 @@ module.exports = function spawnChild (params, callback) {
     let isRunning = true
     try {
       // Signal 0 is a special node construct, see: https://nodejs.org/docs/latest-v14.x/api/process.html#process_process_kill_pid_signal
-      isRunning = process.kill(pid, 0)
+      isRunning = pid === 'init' ? false : process.kill(pid, 0)
     }
     catch (err) {
       isRunning = err.code === 'EPERM'