+++ src/lib/index.js
@@ -1,9 +1,11 @@
+let getFolderSize = require('./get-folder-size')
 let makeRequestId = require('./request-id')
 let runtimeEval = require('./runtime-eval')
 let template = require('./template')
 let userEnvVars = require('./user-env-vars')
 
 module.exports = {
+  getFolderSize,
   makeRequestId,
   runtimeEval,
   template,